requirements.txt
discord.py==2.0.0
notion-client==0.2.0
firebase-admin==5.2.0
python-dotenv==0.20.0
requests==2.26.0
